Fix release clippy gate

+++ crates/kubio-observe/src/observer.rs
@@ -1097,8 +1097,7 @@ impl Observer {
                 vec![AdaptiveReuseBlocker::StaleEvidence],
             );
         }
-        if route.shadow_mismatches as u64 > self.adaptive_reuse.public_object.max_shadow_mismatches
-        {
+        if route.shadow_mismatches > self.adaptive_reuse.public_object.max_shadow_mismatches {
             return ReuseEligibility::blocked(
                 ReuseClass::HardProtected,
                 vec![AdaptiveReuseBlocker::ShadowMismatch],
